+++ src/renderer/components/_molecules/menu-bar/menus/edit.tsx
@@ -1,37 +1,25 @@
 import * as MenuPrimitive from '@radix-ui/react-menubar'
-import { WarningIcon } from '@root/renderer/assets/icons/interface/Warning'
 import { useHandleRemoveTab } from '@root/renderer/hooks'
 import { useOpenPLCStore } from '@root/renderer/store'
 import { i18n } from '@utils/i18n'
 import _ from 'lodash'
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 
-import { Modal, ModalContent } from '../../modal'
 import { MenuClasses } from '../constants'
+import { ConfirmDeletePousModal } from '../modals/delete-confirmation-modal'
 
 export const EditMenu = () => {
   const {
     editor,
-    projectActions: { deletePou },
     workspaceActions: { setModalOpen },
-    flowActions: { removeFlow },
-    libraryActions: { removeUserLibrary },
+    modalActions: { openModal },
+    modals,
   } = useOpenPLCStore()
-  const { handleRemoveTab, selectedTab, setSelectedTab } = useHandleRemoveTab()
+  const { setSelectedTab } = useHandleRemoveTab()
   const { TRIGGER, CONTENT, ITEM, ACCELERATOR, SEPARATOR } = MenuClasses
-  const [modalDeleteOpen, setModalDeleteOpen] = useState(false)
   useEffect(() => {
     setSelectedTab(editor.meta.name)
   }, [editor])
-  const handleDeletePou = () => {
-    handleRemoveTab(selectedTab)
-    deletePou(selectedTab)
-    removeFlow(selectedTab)
-    removeUserLibrary(selectedTab)
-  }
-  const handleCloseModal = () => {
-    setModalDeleteOpen(false)
-  }
 
   const findInProject = () => {
     setModalOpen('findInProject', true)
@@ -91,41 +79,14 @@ export const EditMenu = () => {
             <span>{i18n.t('menu:edit.submenu.selectAll')}</span>
             <span className={ACCELERATOR}>{'Ctrl + A'}</span>
           </MenuPrimitive.Item>
-          <MenuPrimitive.Item className={ITEM} onClick={() => setModalDeleteOpen(true)}>
+          <MenuPrimitive.Item className={ITEM} onClick={() => openModal('confirm-delete-POUs', null)}>
             <span>{i18n.t('menu:edit.submenu.deletePou')}</span>
             <span className={ACCELERATOR}>{''}</span>
           </MenuPrimitive.Item>
         </MenuPrimitive.Content>
       </MenuPrimitive.Portal>
-      {modalDeleteOpen && (
-        <Modal open={modalDeleteOpen} onOpenChange={setModalDeleteOpen}>
-          <ModalContent className='flex max-h-80 w-[300px] select-none flex-col items-center justify-evenly rounded-lg'>
-            <div className='flex select-none flex-col items-center gap-6'>
-              <WarningIcon className='mr-2 mt-2 h-[73px] w-[73px]' />
-              <div>
-                <p className='text-m w-full text-center font-bold text-gray-600 dark:text-neutral-100'>
-                  Are you sure you want to delete this item?
-                </p>
-              </div>
-              <div className='flex w-[200px] flex-col gap-1 space-y-2 text-sm'>
-                <button
-                  onClick={() => {
-                    handleDeletePou()
-                  }}
-                  className='w-full rounded-lg bg-brand px-4 py-2 text-center font-medium text-white'
-                >
-                  Delete
-                </button>
-                <button
-                  onClick={handleCloseModal}
-                  className='w-full rounded-md bg-neutral-100 px-4 py-2 font-medium dark:bg-neutral-850 dark:text-neutral-100 '
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          </ModalContent>
-        </Modal>
+      {modals?.['confirm-delete-POUs']?.open === true && (
+        <ConfirmDeletePousModal isOpen={modals['confirm-delete-POUs'].open} />
       )}
     </MenuPrimitive.Menu>
   )
